+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## added
 
 - more hover documentation for keywords
+- aggregate completions for implicit GROUP BY
 
 ## [0.14.2]